requirements.txt
numpy==2.2.6
pillow==11.2.1
pypng==0.20220715.0
